requirements.txt
ftputil
honcho
hurry.filesize
